Visualizations registry: refactor registry/pluginframework, and promote plugins to class
