+++ lib/galaxy/webapps/galaxy/controllers/visualization.py
@@ -12,7 +12,7 @@ from galaxy.datatypes.interval import Bed
 from galaxy.model.item_attrs import UsesAnnotations, UsesItemRatings
 from galaxy.util.json import loads
 from galaxy.util.sanitize_html import sanitize_html
-from galaxy.visualization import registry
+from galaxy.visualization.plugins import registry
 from galaxy.visualization.data_providers.phyloviz import PhylovizDataProvider
 from galaxy.visualization.data_providers.genome import RawBedDataProvider
 from galaxy.visualization.genomes import decode_dbkey
@@ -25,26 +25,6 @@ from .library import LibraryListGrid
 
 import logging
 log = logging.getLogger( __name__ )
-
-# -- Misc.
-class OpenObject( dict ):
-    #TODO: move to util.data_structures
-    """
-    A dict that allows assignment and attribute retrieval using the dot
-    operator.
-
-    If an attribute isn't contained in the dict `None` is returned (no
-    KeyError).
-    JSON-serializable.
-    """
-    def __getitem__( self, key ):
-        if key not in self:
-            return None
-        return super( OpenObject, self ).__getitem__( key )
-
-    def __getattr__( self, key ):
-        return self.__getitem__( key )
-
 
 
 #
@@ -747,7 +727,7 @@ class VisualizationController( BaseUIController, SharableMixin, UsesVisualizatio
         DEFAULT_VISUALIZATION_NAME = 'Unnamed Visualization'
 
         # post to saved in order to save a visualization
-        #TODO: re-route this one to clear up signature
+        # TODO: re-route this one to clear up signature
         if trans.request.method == 'POST':
             if type is None or config is None:
                 return HTTPBadRequest( 'A visualization type and config are required to save a visualization' )
@@ -755,14 +735,14 @@ class VisualizationController( BaseUIController, SharableMixin, UsesVisualizatio
                 config = loads( config )
             title = title or DEFAULT_VISUALIZATION_NAME
 
-            #TODO: allow saving to (updating) a specific revision - should be part of UsesVisualization
-            #TODO: would be easier if this returned the visualization directly
+            # TODO: allow saving to (updating) a specific revision - should be part of UsesVisualization
+            # TODO: would be easier if this returned the visualization directly
             # check security if posting to existing visualization
             if id is not None:
                 visualization = self.get_visualization( trans, id, check_ownership=True, check_accessible=False )
-                #??: on not owner: error raised, but not returned (status = 200)
+                # ??: on not owner: error raised, but not returned (status = 200)
 
-            #TODO: there's no security check in save visualization (if passed an id)
+            # TODO: there's no security check in save visualization (if passed an id)
             returned = self.save_visualization( trans, config, type, id, title )
 
             # redirect to GET to prevent annoying 'Do you want to post again?' dialog on page reload
@@ -773,8 +753,7 @@ class VisualizationController( BaseUIController, SharableMixin, UsesVisualizatio
             return HTTPBadRequest( 'A valid visualization id is required to load a visualization' )
 
         # render the saved visualization by passing to render, sending latest revision config
-        #TODO: allow loading a specific revision - should be part of UsesVisualization
-        #visualization = self.get_visualization( trans, id, check_ownership=True, check_accessible=False )
+        # TODO: allow loading a specific revision - should be part of UsesVisualization
         visualization = self.get_visualization( trans, id, check_ownership=False, check_accessible=True )
         config = copy.copy( visualization.latest_revision.config )
 
@@ -805,43 +784,13 @@ class VisualizationController( BaseUIController, SharableMixin, UsesVisualizatio
 
         returned = None
         try:
-            # get the config for passing to the template from the kwargs dict, parsed using the plugin's params setting
-            config_from_kwargs = registry.query_dict_to_config( trans, self, visualization_name, kwargs )
-            config.update( config_from_kwargs )
-            config = OpenObject( **config )
-            # further parse config to resources (models, etc.) used in template based on registry config
-            resources = registry.query_dict_to_resources( trans, self, visualization_name, config )
-
-            # if a saved visualization, pass in the encoded visualization id or None if a new render
-            encoded_visualization_id = None
-            if visualization:
-                encoded_visualization_id = trans.security.encode_id( visualization.id )
-
-            visualization_display_name = plugin.config[ 'name' ]
-            title = visualization.latest_revision.title if visualization else kwargs.get( 'title', None )
-
-            # look up template and render
-            template_path = plugin.config[ 'template' ]
-            returned = registry.fill_template( trans, plugin, template_path,
-                visualization_name=visualization_name,
-                visualization_display_name=visualization_display_name,
-                title=title,
-                saved_visualization=visualization,
-                visualization_id=encoded_visualization_id,
-                embedded=embedded,
-                #NOTE: passing *unparsed* kwargs as query
-                query=kwargs,
-                #NOTE: vars is a dictionary for shared data in the template
-                #   this feels hacky to me but it's what mako recommends:
-                #   http://docs.makotemplates.org/en/latest/runtime.html
-                vars={},
-                config=config,
-                **resources
-            )
+            returned = plugin.render( controller=self, trans=trans,
+                visualization=visualization, config=config, embedded=embedded, **kwargs )
 
         except Exception, exception:
             log.exception( 'error rendering visualization (%s): %s', visualization_name, str( exception ) )
-            if trans.debug: raise
+            if trans.debug:
+                raise
             returned = trans.show_error_message(
                 "There was an error rendering the visualization. " +
                 "Contact your Galaxy administrator if the problem persists." +
